requirements.txt
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Cors==3.0.10
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.3.0
qrcode==7.3.1
six==1.16.0
uuid==1.30
Werkzeug==2.2.2
zipp==3.11.0
